requirements.txt
flask==3.0.2
gunicorn==21.2.0
aiohttp==3.9.3
asyncio==3.4.3
python-dotenv==1.0.0
anthropic==0.7.8 